:ok_hand: Wrapped save in atomic transaction
As correctly noted by Joeri, if the save is not part of a bigger/outer
atomic transaction block, the on_commit handler is invoked immediately,
which runs the sync code instantly BEFORE the relevant object is
actually persisted in the database. This means that the authorization
will not correctly be added as would be expected.

Adding the atomic transaction blocks ensures this only happens after the
relevant object has been persisted to the database.