:construction_worker: Make Newman/Docker stages conditional
Via #379 it became clear that forks (untrusted) do not have access
to the necessary environment variables to:
* run Newman tests
* publish docker images

This should disable the relevant stages for forks, keeping the Django
tests and code style checks stage. Newman tests and docker builds/push
will happen once the branch is merged into master.

Note that this introduces a small regression risk - changes could be
introduced that break the API tests from the ATP and we'd only
realize _after_ the merge to master.