Added CMIS integration for the documents API backend
CMIS 1.1 protocol is used to talk to CMIS servers such as Alfresco.

Note that this is one big commit for the entire integration. We're
sorry, but the feature was massive and the commits got interwoven with
merges back from master. So, the entire feature is now a single commit.

The original PR can be found here:
https://github.com/open-zaak/open-zaak/pull/540

The branch name is 'non-intrusive_cmis-integration'.