:art: Clean up tests and fix implementation problems
Removed all the added /unlock calls for documents, which weren't present
in the original test cases and have distinctly different behaviour.

The CMIS queryset implementation now does as much as possible in the
CMIS query (i.e. the CMIS query mechanism as part of the CMIS
standard), and what's left over is handled in Python on the results
from the CMIS query.

Note that the way we deal with versions is substantially different
from the initial implementation. The Documenten API has monotonically
increasing integer version numbers, while the CMIS standard prescribes
the usage of cmis:versionLabel _without_ defining an exact format,
and the repository does not have a MUST requirement to bump versions.
Alfresco follows a major.minor format for this, where property updates
bump the major version (as it seems). In short - under the hood
CMIS actions happen that make the version progression faster than the
one intended in the API. Therefore, a property drc:document__versie
is defined in the content model which tracks the version format as
required for the API spec. Version numbers are no longer derived from
cmis:versionLabel. This also means that multiple documents can exist
in the CMIS repository with the combination (uuid, versie) - which is
fine as long as we make sure to grab the latest CMIS version for
each combination (based on cmis:creationDate).