:fire: Fixes #860 -- remove autorisatie-inlines for applicaties
The 'manage authorizations' view presents a condensed, aggregated view
of the different authorizations for a given API application.

The regular admin interface displayed all authorizations as inline with
the application, but because of the amount of possible combinations
(different zaaktypen, documenttypen, besluittypen... and permissions),
this inline can quite easily exceed 10.000 records.

Once that number is exceeded, Django blocks the request because
`DATA_UPLOAD_MAX_NUMBER_FIELDS` is exceeded. Removing the inlines
fixes this.