:sparkles: Fixes #1085 -- perform semantic valdation on zaaktype publish
* Validate that at least one roltype is defined
* Validate that at least one resultaattype is defined
* Validate that at least two statustypen are defined

Closing a zaak requires a resultaat, which requires a resultaattype.
Starting a zaak requires an initial statustype. Closing a zaak
also requires a statustype with the highest 'volgnummer'. Because
a resultaat must be set between opening and closing a zaak,
there must be two distinct statustypen.

Roltypen are required to be able to register who does what within the
zaak (most notably, initiator). However, initiator does not appear
to be required from RGBZ 2.0/ImZTC.