:whale: Reduce docker build context and image size
This excludes tests, cached bytecode and test assets from ending up in the final container
image.
This code is never meant to be run or used in production scenario's,
so we remove dead weight. A nice benefit is that the container image
should be a bit smaller in the end.