:white_check_mark: [#669] -- define end-to-end tests for setup_configuration command
The setup_configuration management command needs to operate in a number
of different modes:

* 100% headless without any user interaction
* in interactive mode where missing data is prompted for

Various combinations are possible, and certain configuration aspects
should be able to be skipped without wiping configuration.

The command must also correctly offer to self-test the configuration
if not provided explicitly.

Finally, it should be able to chain this command to send an actual
test notification to the Notifications API for a complete end-to-end
check of valid configuration.