:card_file_box: [#1271] -- set up migrations for multi-table inheritance
The Zaak model ideally is multi-table inheritance so it shares the
same primary key/ID with the ZaakIdentificatie model. An extra
advantage is that the existing serializer/viewset/admin code keeps
working as is, and the abstraction between the different tables is
handled by Django.

However, migrating an existing model to be multi-table inheritance is
quite an undertaking, which is done in these migration files for this
commit. The essence is:

* Every model with FK to Zaak needs to temporarily drop the constraints,
  we do this by converting into a regular IntegerField and then
  swapping back.
* Fields moved from Zaak to ZaakIdentificatie need to be renamed (but
  for now we keep the existing columns so we can reverse the
  migrations, otherwise we get IntegrityError on NULL values in NOT NULL
  columns)
* The virtual fields are added in pure state operations, as they don't
  have a backing database field. The derived constraint updates are
  handled similarly

These migrations are generated by tweaking the models and commenting
out a lot of code to make drf/django not trip about the unexpected
integer types (instead of FKs). Then they were manually re-organized
and adapted, after which the temporary code changes were reverted
and/or uncommented to what it should be. The generated migrations
from that re-instate the FK constraints.

Note that zaken 0017 migration takes care of copying over the exact
same IDs (and sequences) from the Zaak -> ZaakIdentificatie tabel,
with a table lock that prevents writes to the Zaak table while the
migration is running. This is crucial to be able to update the FK
constraints to the new table instead of the ('deleted') Zaak.id
column.