:recycle: [#1733] Refactor polymorphic serializer structure
The polymorphic serializers for NP, NNP, Vestiging... etc. were
automatically lumped in a nested container 'betrokkene_identificatie',
however this won't work if we add additional fields to the involved
models that should be exposed on a higher level.

The NatuurlijkPersoon model/serializer setup prototypes what this can
look like, and the remaining serializers have been updated so that the
original API schema is retained.